fix(docs): address plugin preview QR feedback
