+++ apps/shared/pluginPreviewQrs.ts
@@ -1,0 +1,791 @@
+export interface PluginPreviewQr {
+  appName: string
+  appId: string
+  channelName: string
+  channelId: number
+  version: string
+  webUrl: string
+  qr: string
+}
+
+interface PluginPreviewQrSource {
+  appName: string
+  appId: string
+  channelName: string
+  channelId: number
+  version: string
+}
+
+const buildPreviewUrl = (appId: string, channelId: number): string => `https://c${channelId}-${appId.replaceAll('.', '-0')}.preview.capgo.app/`
+const buildQrPath = (appId: string, channelName: string, channelId: number): string => `/plugins/preview-qrs/${appId.replaceAll('.', '-')}-${channelName}-${channelId}.webp`
+
+const pluginPreviewQrSources: Record<string, PluginPreviewQrSource> = {
+  'capacitor-accelerometer': {
+    appName: 'Accelerometer Example',
+    appId: 'app.capgo.accelerometer',
+    channelName: 'production',
+    channelId: 38840,
+    version: '1.0.0',
+  },
+  'capacitor-admob': {
+    appName: 'Admob Example',
+    appId: 'app.capgo.admob',
+    channelName: 'production',
+    channelId: 38920,
+    version: '1.0.0',
+  },
+  'capacitor-alarm': {
+    appName: 'Alarm Example',
+    appId: 'app.capgo.alarm',
+    channelName: 'production',
+    channelId: 38837,
+    version: '1.0.0',
+  },
+  'capacitor-android-age-signals': {
+    appName: 'Age Signals Example',
+    appId: 'app.capgo.androidagesignals.example',
+    channelName: 'production',
+    channelId: 38835,
+    version: '1.0.0',
+  },
+  'capacitor-android-inline-install': {
+    appName: 'Android Inline Install Example',
+    appId: 'app.capgo.android.inline.install',
+    channelName: 'production',
+    channelId: 38838,
+    version: '1.0.0',
+  },
+  'capacitor-android-sms-retriever': {
+    appName: 'Plugin Template Example',
+    appId: 'app.capgo.androidsmsretriever.example',
+    channelName: 'production',
+    channelId: 38895,
+    version: '1.0.0',
+  },
+  'capacitor-android-usagestatsmanager': {
+    appName: 'Android Usage Stats Manager Example',
+    appId: 'app.capgo.android.usagestatsmanager',
+    channelName: 'production',
+    channelId: 38850,
+    version: '1.0.0',
+  },
+  'capacitor-app-attest': {
+    appName: 'App Attest Example',
+    appId: 'app.capgo.app.attest',
+    channelName: 'production',
+    channelId: 38915,
+    version: '1.0.0',
+  },
+  'capacitor-appinsights': {
+    appName: 'App Insights Example',
+    appId: 'app.capgo.appinsights',
+    channelName: 'production',
+    channelId: 38834,
+    version: '0.0.0',
+  },
+  'capacitor-appsflyer': {
+    appName: 'AppsFlyer Example',
+    appId: 'app.capgo.appsflyer.example',
+    channelName: 'production',
+    channelId: 38839,
+    version: '1.0.0',
+  },
+  'capacitor-audio-recorder': {
+    appName: 'Audio Recorder Example',
+    appId: 'app.capgo.audiorecorder',
+    channelName: 'production',
+    channelId: 38925,
+    version: '1.0.0',
+  },
+  'capacitor-auto': {
+    appName: 'Plugin Template Example',
+    appId: 'app.capgo.auto.example',
+    channelName: 'production',
+    channelId: 38873,
+    version: '1.0.0',
+  },
+  'capacitor-autofill-save-password': {
+    appName: 'Autofill Save Password Example',
+    appId: 'app.capgo.autofill.save.password.example',
+    channelName: 'production',
+    channelId: 38937,
+    version: '1.0.0',
+  },
+  'capacitor-background-geolocation': {
+    appName: 'Background Geolocation Example',
+    appId: 'com.capgo.backgroundgeolocation.example',
+    channelName: 'production',
+    channelId: 38888,
+    version: '1.0.0',
+  },
+  'capacitor-background-task': {
+    appName: 'Background Task Example',
+    appId: 'app.capgo.backgroundtask.example',
+    channelName: 'production',
+    channelId: 38864,
+    version: '1.0.0',
+  },
+  'capacitor-barometer': {
+    appName: 'Barometer Example',
+    appId: 'app.capgo.barometer',
+    channelName: 'production',
+    channelId: 38939,
+    version: '1.0.0',
+  },
+  'capacitor-bluetooth-low-energy': {
+    appName: 'BLE Example',
+    appId: 'ee.forgr.plugin.bluetoothlowenergy.example',
+    channelName: 'production',
+    channelId: 38887,
+    version: '0.0.0',
+  },
+  'capacitor-brightness': {
+    appName: 'Brightness Example',
+    appId: 'app.capgo.brightness',
+    channelName: 'production',
+    channelId: 38979,
+    version: '1.0.0',
+  },
+  'capacitor-calendar': {
+    appName: 'Capgo Calendar',
+    appId: 'app.capgo.calendar.example',
+    channelName: 'production',
+    channelId: 37574,
+    version: '8.0.7',
+  },
+  'capacitor-camera-preview': {
+    appName: 'Camera Preview Example',
+    appId: 'app.capgo.camera.preview',
+    channelName: 'production',
+    channelId: 38918,
+    version: '0.0.1',
+  },
+  'capacitor-compass': {
+    appName: 'Compass Example',
+    appId: 'app.capgo.compass',
+    channelName: 'production',
+    channelId: 38879,
+    version: '1.0.0',
+  },
+  'capacitor-contacts': {
+    appName: 'Contacts Example',
+    appId: 'app.capgo.contacts',
+    channelName: 'production',
+    channelId: 38886,
+    version: '1.0.0',
+  },
+  'capacitor-contentsquare': {
+    appName: 'Contentsquare Example',
+    appId: 'app.capgo.contentsquare.example',
+    channelName: 'production',
+    channelId: 38912,
+    version: '1.0.0',
+  },
+  'capacitor-crisp': {
+    appName: 'Crisp Example',
+    appId: 'app.capgo.crisp',
+    channelName: 'production',
+    channelId: 38862,
+    version: '1.0.0',
+  },
+  'capacitor-data-storage-sqlite': {
+    appName: 'Data Storage SQLite Example',
+    appId: 'app.capgo.data.storage.sqlite',
+    channelName: 'production',
+    channelId: 38931,
+    version: '1.0.0',
+  },
+  'capacitor-date-picker': {
+    appName: 'Capgo Date Picker',
+    appId: 'app.capgo.datepicker.example',
+    channelName: 'production',
+    channelId: 37575,
+    version: '8.1.1-22dd072',
+  },
+  'capacitor-device-info': {
+    appName: 'Device Info Example',
+    appId: 'app.capgo.deviceinfo.example',
+    channelName: 'production',
+    channelId: 38927,
+    version: '1.0.0',
+  },
+  'capacitor-document-scanner': {
+    appName: 'Document Scanner Example',
+    appId: 'app.capgo.document.scanner',
+    channelName: 'production',
+    channelId: 38891,
+    version: '1.0.0',
+  },
+  'capacitor-downloader': {
+    appName: 'Downloader Example',
+    appId: 'app.capgo.downloader',
+    channelName: 'production',
+    channelId: 38936,
+    version: '1.0.0',
+  },
+  'capacitor-env': {
+    appName: 'Environment Example',
+    appId: 'app.capgo.env',
+    channelName: 'production',
+    channelId: 38942,
+    version: '1.0.0',
+  },
+  'capacitor-facebook-analytics': {
+    appName: 'Facebook Analytics Example',
+    appId: 'app.capgo.facebookanalytics.example',
+    channelName: 'production',
+    channelId: 38857,
+    version: '1.0.0',
+  },
+  'capacitor-fast-sql': {
+    appName: 'FastSQL Example',
+    appId: 'com.capgo.fastsql.example',
+    channelName: 'production',
+    channelId: 38907,
+    version: '1.0.0',
+  },
+  'capacitor-file': {
+    appName: 'File Example',
+    appId: 'app.capgo.file',
+    channelName: 'production',
+    channelId: 38831,
+    version: '1.0.0',
+  },
+  'capacitor-file-compressor': {
+    appName: 'File Compressor Example',
+    appId: 'app.capgo.filecompressor',
+    channelName: 'production',
+    channelId: 38922,
+    version: '1.0.0',
+  },
+  'capacitor-file-sharer': {
+    appName: 'Plugin Template Example',
+    appId: 'app.capgo.filesharer.example',
+    channelName: 'production',
+    channelId: 38905,
+    version: '1.0.0',
+  },
+  'capacitor-flash': {
+    appName: 'Flash Example',
+    appId: 'app.capgo.flash',
+    channelName: 'production',
+    channelId: 38896,
+    version: '1.0.0',
+  },
+  'capacitor-gtm': {
+    appName: 'GTM Example',
+    appId: 'app.capgo.gtm',
+    channelName: 'production',
+    channelId: 38855,
+    version: '1.0.0',
+  },
+  'capacitor-health': {
+    appName: 'Health Example',
+    appId: 'app.capgo.health',
+    channelName: 'production',
+    channelId: 38903,
+    version: '1.0.0',
+  },
+  'capacitor-home-indicator': {
+    appName: 'Home Indicator Example',
+    appId: 'app.capgo.home.indicator.example',
+    channelName: 'production',
+    channelId: 38900,
+    version: '1.0.0',
+  },
+  'capacitor-ibeacon': {
+    appName: 'iBeacon Example',
+    appId: 'app.capgo.ibeacon',
+    channelName: 'production',
+    channelId: 38851,
+    version: '1.0.0',
+  },
+  'capacitor-inappbrowser': {
+    appName: 'Inappbrowser Example',
+    appId: 'app.capgo.inappbrowser',
+    channelName: 'production',
+    channelId: 38930,
+    version: '1.0.0',
+  },
+  'capacitor-incoming-call-kit': {
+    appName: 'Privacy Screen Example',
+    appId: 'app.capgo.incomingcallkit.example',
+    channelName: 'production',
+    channelId: 38841,
+    version: '1.0.0',
+  },
+  'capacitor-install-referrer': {
+    appName: 'Install Referrer Example',
+    appId: 'app.capgo.installreferrer.example',
+    channelName: 'production',
+    channelId: 38832,
+    version: '1.0.0',
+  },
+  'capacitor-intent-launcher': {
+    appName: 'Intent Launcher Example',
+    appId: 'app.capgo.intentlauncher',
+    channelName: 'production',
+    channelId: 38904,
+    version: '1.0.0',
+  },
+  'capacitor-is-root': {
+    appName: 'Is Root Example',
+    appId: 'app.capgo.is.root',
+    channelName: 'production',
+    channelId: 38902,
+    version: '1.0.0',
+  },
+  'capacitor-ivs-player': {
+    appName: 'IVS Player Example',
+    appId: 'app.capgo.ivs.player.example',
+    channelName: 'production',
+    channelId: 38871,
+    version: '1.0.0',
+  },
+  'capacitor-jw-player': {
+    appName: 'Jw Player Example',
+    appId: 'app.capgo.jw.player',
+    channelName: 'production',
+    channelId: 38964,
+    version: '1.0.0',
+  },
+  'capacitor-keep-awake': {
+    appName: 'Keep Awake Example',
+    appId: 'app.capgo.keepawake',
+    channelName: 'production',
+    channelId: 38919,
+    version: '1.0.0',
+  },
+  'capacitor-launch-navigator': {
+    appName: 'Launch Navigator Example',
+    appId: 'app.capgo.launch.navigator',
+    channelName: 'production',
+    channelId: 38880,
+    version: '1.0.0',
+  },
+  'capacitor-light-sensor': {
+    appName: 'Light Sensor Example',
+    appId: 'app.capgo.lightsensor',
+    channelName: 'production',
+    channelId: 38882,
+    version: '0.0.1',
+  },
+  'capacitor-live-reload': {
+    appName: 'Live Reload Example',
+    appId: 'app.capgo.livereload',
+    channelName: 'production',
+    channelId: 38897,
+    version: '1.0.0',
+  },
+  'capacitor-llm': {
+    appName: 'LLM Example',
+    appId: 'app.capgo.llm',
+    channelName: 'production',
+    channelId: 38940,
+    version: '0.0.1',
+  },
+  'capacitor-media-session': {
+    appName: 'Media Session Example',
+    appId: 'app.capgo.media.session',
+    channelName: 'production',
+    channelId: 38941,
+    version: '1.0.0',
+  },
+  'capacitor-mqtt': {
+    appName: 'MQTT Example',
+    appId: 'app.capgo.mqtt.example',
+    channelName: 'production',
+    channelId: 38858,
+    version: '1.0.0',
+  },
+  'capacitor-mute': {
+    appName: 'Mute Example',
+    appId: 'app.capgo.mute',
+    channelName: 'production',
+    channelId: 38833,
+    version: '1.0.0',
+  },
+  'capacitor-mux-player': {
+    appName: 'Mux Player Example',
+    appId: 'app.capgo.mux.player',
+    channelName: 'production',
+    channelId: 38898,
+    version: '1.0.0',
+  },
+  'capacitor-native-biometric': {
+    appName: 'example-app',
+    appId: 'app.capgo.native.biometric',
+    channelName: 'production',
+    channelId: 38893,
+    version: '1.0.0',
+  },
+  'capacitor-native-loader': {
+    appName: 'Native Loader Example',
+    appId: 'app.capgo.nativeloader.example',
+    channelName: 'production',
+    channelId: 38926,
+    version: '1.0.0',
+  },
+  'capacitor-native-market': {
+    appName: 'Native Market Example',
+    appId: 'app.capgo.nativemarket',
+    channelName: 'production',
+    channelId: 38911,
+    version: '1.0.0',
+  },
+  'capacitor-native-navigation': {
+    appName: 'Native Navigation Example',
+    appId: 'app.capgo.native.navigation.example',
+    channelName: 'production',
+    channelId: 38546,
+    version: '1.0.0',
+  },
+  'capacitor-native-purchases': {
+    appName: 'Native Purchases Example',
+    appId: 'app.capgo.nativepurchases',
+    channelName: 'production',
+    channelId: 38889,
+    version: '1.0.0',
+  },
+  'capacitor-nativegeocoder': {
+    appName: 'Native Geocoder Example',
+    appId: 'app.capgo.nativegeocoder',
+    channelName: 'production',
+    channelId: 38842,
+    version: '1.0.0',
+  },
+  'capacitor-navigation-bar': {
+    appName: 'Navigation Bar Example',
+    appId: 'app.capgo.navigation.bar',
+    channelName: 'production',
+    channelId: 38854,
+    version: '0.0.1',
+  },
+  'capacitor-network-diagnostics': {
+    appName: 'Network Diagnostics Example',
+    appId: 'app.capgo.networkdiagnostics.example',
+    channelName: 'production',
+    channelId: 38869,
+    version: '1.0.0',
+  },
+  'capacitor-nfc': {
+    appName: 'NFC Example',
+    appId: 'app.capgo.nfc.example',
+    channelName: 'production',
+    channelId: 38932,
+    version: '1.0.0',
+  },
+  'capacitor-passkey': {
+    appName: 'Capgo Passkey Demo',
+    appId: 'app.capgo.passkey.example',
+    channelName: 'production',
+    channelId: 38853,
+    version: '1.0.0',
+  },
+  'capacitor-pay': {
+    appName: 'Pay Example',
+    appId: 'app.capgo.pay',
+    channelName: 'production',
+    channelId: 38870,
+    version: '1.0.0',
+  },
+  'capacitor-pdf-generator': {
+    appName: 'PDF Generator Example',
+    appId: 'app.capgo.pdfgenerator',
+    channelName: 'production',
+    channelId: 38861,
+    version: '1.0.0',
+  },
+  'capacitor-pedometer': {
+    appName: 'Pedometer Example',
+    appId: 'app.capgo.pedometer',
+    channelName: 'production',
+    channelId: 38909,
+    version: '1.0.0',
+  },
+  'capacitor-persistent-account': {
+    appName: 'Persistent Account Example',
+    appId: 'app.capgo.persistent.account',
+    channelName: 'production',
+    channelId: 38849,
+    version: '1.0.0',
+  },
+  'capacitor-photo-library': {
+    appName: 'Photo Library Example',
+    appId: 'app.capgo.photo.library',
+    channelName: 'production',
+    channelId: 38894,
+    version: '1.0.0',
+  },
+  'capacitor-pretty-toast': {
+    appName: 'Pretty Toast',
+    appId: 'app.capgo.pretty.toast',
+    channelName: 'production',
+    channelId: 37182,
+    version: '8.1.10',
+  },
+  'capacitor-printer': {
+    appName: 'CapacitorPrinterExample',
+    appId: 'com.capgo.printer.example',
+    channelName: 'production',
+    channelId: 38885,
+    version: '1.0.0',
+  },
+  'capacitor-privacy-screen': {
+    appName: 'Privacy Screen Example',
+    appId: 'app.capgo.privacyscreen.example',
+    channelName: 'production',
+    channelId: 38916,
+    version: '1.0.0',
+  },
+  'capacitor-proximity': {
+    appName: 'Capgo Proximity Demo',
+    appId: 'app.capgo.proximity.example',
+    channelName: 'production',
+    channelId: 38901,
+    version: '1.0.0',
+  },
+  'capacitor-realtimekit': {
+    appName: 'RealtimeKit Example',
+    appId: 'app.capgo.realtimekit',
+    channelName: 'production',
+    channelId: 38848,
+    version: '1.0.0',
+  },
+  'capacitor-recaptcha': {
+    appName: 'Recaptcha Example',
+    appId: 'app.capgo.recaptcha.example',
+    channelName: 'production',
+    channelId: 38881,
+    version: '1.0.0',
+  },
+  'capacitor-ricoh360-camera-plugin': {
+    appName: 'Ricoh360 Camera Plugin Example',
+    appId: 'app.capgo.ricoh360.camera.plugin',
+    channelName: 'production',
+    channelId: 38868,
+    version: '1.0.0',
+  },
+  'capacitor-rudderstack': {
+    appName: 'RudderStack Example',
+    appId: 'app.capgo.rudderstack.example',
+    channelName: 'production',
+    channelId: 38906,
+    version: '1.0.0',
+  },
+  'capacitor-screen-orientation': {
+    appName: 'Screen Orientation Example',
+    appId: 'app.capgo.screenorientation',
+    channelName: 'production',
+    channelId: 38934,
+    version: '1.0.0',
+  },
+  'capacitor-screen-recorder': {
+    appName: 'Screen Recorder Example',
+    appId: 'app.capgo.screen.recorder',
+    channelName: 'production',
+    channelId: 38884,
+    version: '1.0.0',
+  },
+  'capacitor-shake': {
+    appName: 'Shake Example',
+    appId: 'app.capgo.shake',
+    channelName: 'production',
+    channelId: 38938,
+    version: '1.0.0',
+  },
+  'capacitor-share-target': {
+    appName: 'Share Target Example',
+    appId: 'app.capgo.sharetarget',
+    channelName: 'production',
+    channelId: 38928,
+    version: '1.0.0',
+  },
+  'capacitor-sheets': {
+    appName: 'Capgo Sheets',
+    appId: 'app.capgo.capacitor.sheets',
+    channelName: 'production',
+    channelId: 37983,
+    version: '8.0.17',
+  },
+  'capacitor-sim': {
+    appName: 'Sim Example',
+    appId: 'app.capgo.sim',
+    channelName: 'production',
+    channelId: 38910,
+    version: '1.0.0',
+  },
+  'capacitor-speech-recognition': {
+    appName: 'Speech Recognition Example',
+    appId: 'app.capgo.speechrecognition',
+    channelName: 'production',
+    channelId: 38935,
+    version: '1.0.0',
+  },
+  'capacitor-speech-synthesis': {
+    appName: 'Speech Synthesis Example',
+    appId: 'app.capgo.speechsynthesis',
+    channelName: 'production',
+    channelId: 38872,
+    version: '1.0.0',
+  },
+  'capacitor-ssl-pinning': {
+    appName: 'SSL Pinning Example',
+    appId: 'app.capgo.sslpinning.example',
+    channelName: 'production',
+    channelId: 38923,
+    version: '1.0.0',
+  },
+  'capacitor-supabase': {
+    appName: 'Supabase Example',
+    appId: 'app.capgo.supabase.example',
+    channelName: 'production',
+    channelId: 38878,
+    version: '1.0.0',
+  },
+  'capacitor-textinteraction': {
+    appName: 'Text Interaction Example',
+    appId: 'app.capgo.textinteraction',
+    channelName: 'production',
+    channelId: 38917,
+    version: '1.0.0',
+  },
+  'capacitor-transitions': {
+    appName: 'Capgo Transitions',
+    appId: 'app.capgo.capacitor.transitions',
+    channelName: 'production',
+    channelId: 37981,
+    version: '8.1.9',
+  },
+  'capacitor-twilio-voice': {
+    appName: 'Twilio Voice Example',
+    appId: 'app.capgo.twilio.voice',
+    channelName: 'production',
+    channelId: 38883,
+    version: '1.0.0',
+  },
+  'capacitor-updater': {
+    appName: 'Updater Example',
+    appId: 'app.capgo.updater',
+    channelName: 'production',
+    channelId: 39087,
+    version: '1.0.0',
+  },
+  'capacitor-uploader': {
+    appName: 'Uploader Example',
+    appId: 'app.capgo.uploader',
+    channelName: 'production',
+    channelId: 38929,
+    version: '1.0.0',
+  },
+  'capacitor-verisoul': {
+    appName: 'Verisoul Example',
+    appId: 'app.capgo.verisoul.example',
+    channelName: 'production',
+    channelId: 38899,
+    version: '1.0.0',
+  },
+  'capacitor-video-player': {
+    appName: 'Video Player Example',
+    appId: 'app.capgo.video.player',
+    channelName: 'production',
+    channelId: 38877,
+    version: '1.0.0',
+  },
+  'capacitor-video-thumbnails': {
+    appName: 'Video Thumbnails Example',
+    appId: 'app.capgo.videothumbnails',
+    channelName: 'production',
+    channelId: 38913,
+    version: '1.0.0',
+  },
+  'capacitor-volume-buttons': {
+    appName: 'Volume Buttons Example',
+    appId: 'app.capgo.volume.buttons',
+    channelName: 'production',
+    channelId: 38847,
+    version: '1.0.0',
+  },
+  'capacitor-watch': {
+    appName: 'Watch Example',
+    appId: 'app.capgo.watch',
+    channelName: 'production',
+    channelId: 38892,
+    version: '1.0.0',
+  },
+  'capacitor-webview-crash': {
+    appName: 'WebView Crash Demo',
+    appId: 'app.capgo.webviewcrash.example',
+    channelName: 'production',
+    channelId: 38890,
+    version: '1.0.0',
+  },
+  'capacitor-webview-guardian': {
+    appName: 'Webview Guardian Example',
+    appId: 'app.capgo.webviewguardian.example',
+    channelName: 'production',
+    channelId: 38924,
+    version: '1.0.0',
+  },
+  'capacitor-webview-version-checker': {
+    appName: 'WebView Version Checker Example',
+    appId: 'app.capgo.webviewversionchecker.example',
+    channelName: 'production',
+    channelId: 38948,
+    version: '1.0.0',
+  },
+  'capacitor-wechat': {
+    appName: 'WeChat Example',
+    appId: 'app.capgo.wechat',
+    channelName: 'production',
+    channelId: 38875,
+    version: '1.0.0',
+  },
+  'capacitor-widget-kit': {
+    appName: 'Widget Kit Example',
+    appId: 'app.capgo.widgetkit.exampleapp',
+    channelName: 'production',
+    channelId: 38836,
+    version: '1.0.0',
+  },
+  'capacitor-wifi': {
+    appName: 'WiFi Example',
+    appId: 'app.capgo.wifi',
+    channelName: 'production',
+    channelId: 38856,
+    version: '1.0.0',
+  },
+  'capacitor-youtube-player': {
+    appName: 'YouTube Player Example',
+    appId: 'app.capgo.youtube.player',
+    channelName: 'production',
+    channelId: 38846,
+    version: '1.0.0',
+  },
+  'capacitor-zebra-datawedge': {
+    appName: 'Zebra DataWedge Example',
+    appId: 'app.capgo.zebradatawedge.example',
+    channelName: 'production',
+    channelId: 38914,
+    version: '1.0.0',
+  },
+  'capacitor-zip': {
+    appName: 'Zip Example',
+    appId: 'app.capgo.zip',
+    channelName: 'production',
+    channelId: 38863,
+    version: '1.0.0',
+  },
+}
+
+export const pluginPreviewQrs = Object.entries(pluginPreviewQrSources).reduce<Record<string, PluginPreviewQr>>((acc, [slug, preview]) => {
+  acc[slug] = {
+    ...preview,
+    webUrl: buildPreviewUrl(preview.appId, preview.channelId),
+    qr: buildQrPath(preview.appId, preview.channelName, preview.channelId),
+  }
+
+  return acc
+}, {})
+
+export const getPluginPreviewQr = (slug?: string): PluginPreviewQr | undefined => (slug ? pluginPreviewQrs[slug] : undefined)
