+++ apps/web/src/components/PluginPreviewQrCard.astro
@@ -22,7 +22,10 @@ const imageAlt = `${title} plugin preview QR code`
     href={preview.webUrl}
     target="_blank"
     rel="noopener noreferrer"
-    class:list={['block shrink-0 overflow-hidden rounded-xl border border-white/15 bg-white p-2 transition hover:border-cyan-200/60', compact ? 'w-28' : 'mx-auto w-40']}
+    class:list={[
+      'block shrink-0 overflow-hidden rounded-xl border border-white/15 bg-white p-2 transition hover:border-cyan-200/60 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-400',
+      compact ? 'w-28' : 'mx-auto w-40',
+    ]}
   >
     <img src={preview.qr} alt={imageAlt} loading="lazy" decoding="async" class="h-auto w-full" />
   </a>
@@ -34,7 +37,7 @@ const imageAlt = `${title} plugin preview QR code`
       href={preview.webUrl}
       target="_blank"
       rel="noopener noreferrer"
-      class="mt-3 inline-flex cursor-pointer items-center gap-2 text-sm font-semibold text-cyan-100 transition hover:text-white"
+      class="mt-3 inline-flex cursor-pointer items-center gap-2 text-sm font-semibold text-cyan-100 transition hover:text-white focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-400"
     >
       <span>Open preview</span>
       <AstroIcon name="heroicons:arrow-up-right-solid" class="size-4" aria-hidden="true" />
