+++ apps/web/src/config/pluginPreviewQrs.ts
@@ -1,986 +1,1 @@
-export interface PluginPreviewQr {
-  appName: string
-  appId: string
-  channelName: string
-  channelId: number
-  version: string
-  webUrl: string
-  qr: string
-}
-
-export const pluginPreviewQrs: Record<string, PluginPreviewQr> = {
-  'capacitor-accelerometer': {
-    appName: 'Accelerometer Example',
-    appId: 'app.capgo.accelerometer',
-    channelName: 'production',
-    channelId: 38840,
-    version: '1.0.0',
-    webUrl: 'https://c38840-app-0capgo-0accelerometer.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-accelerometer-production-38840.webp',
-  },
-  'capacitor-admob': {
-    appName: 'Admob Example',
-    appId: 'app.capgo.admob',
-    channelName: 'production',
-    channelId: 38920,
-    version: '1.0.0',
-    webUrl: 'https://c38920-app-0capgo-0admob.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-admob-production-38920.webp',
-  },
-  'capacitor-alarm': {
-    appName: 'Alarm Example',
-    appId: 'app.capgo.alarm',
-    channelName: 'production',
-    channelId: 38837,
-    version: '1.0.0',
-    webUrl: 'https://c38837-app-0capgo-0alarm.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-alarm-production-38837.webp',
-  },
-  'capacitor-android-age-signals': {
-    appName: 'Age Signals Example',
-    appId: 'app.capgo.androidagesignals.example',
-    channelName: 'production',
-    channelId: 38835,
-    version: '1.0.0',
-    webUrl: 'https://c38835-app-0capgo-0androidagesignals-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-androidagesignals-example-production-38835.webp',
-  },
-  'capacitor-android-inline-install': {
-    appName: 'Android Inline Install Example',
-    appId: 'app.capgo.android.inline.install',
-    channelName: 'production',
-    channelId: 38838,
-    version: '1.0.0',
-    webUrl: 'https://c38838-app-0capgo-0android-0inline-0install.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-android-inline-install-production-38838.webp',
-  },
-  'capacitor-android-sms-retriever': {
-    appName: 'Plugin Template Example',
-    appId: 'app.capgo.androidsmsretriever.example',
-    channelName: 'production',
-    channelId: 38895,
-    version: '1.0.0',
-    webUrl: 'https://c38895-app-0capgo-0androidsmsretriever-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-androidsmsretriever-example-production-38895.webp',
-  },
-  'capacitor-android-usagestatsmanager': {
-    appName: 'Android Usage Stats Manager Example',
-    appId: 'app.capgo.android.usagestatsmanager',
-    channelName: 'production',
-    channelId: 38850,
-    version: '1.0.0',
-    webUrl: 'https://c38850-app-0capgo-0android-0usagestatsmanager.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-android-usagestatsmanager-production-38850.webp',
-  },
-  'capacitor-app-attest': {
-    appName: 'App Attest Example',
-    appId: 'app.capgo.app.attest',
-    channelName: 'production',
-    channelId: 38915,
-    version: '1.0.0',
-    webUrl: 'https://c38915-app-0capgo-0app-0attest.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-app-attest-production-38915.webp',
-  },
-  'capacitor-appinsights': {
-    appName: 'App Insights Example',
-    appId: 'app.capgo.appinsights',
-    channelName: 'production',
-    channelId: 38834,
-    version: '0.0.0',
-    webUrl: 'https://c38834-app-0capgo-0appinsights.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-appinsights-production-38834.webp',
-  },
-  'capacitor-appsflyer': {
-    appName: 'AppsFlyer Example',
-    appId: 'app.capgo.appsflyer.example',
-    channelName: 'production',
-    channelId: 38839,
-    version: '1.0.0',
-    webUrl: 'https://c38839-app-0capgo-0appsflyer-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-appsflyer-example-production-38839.webp',
-  },
-  'capacitor-audio-recorder': {
-    appName: 'Audio Recorder Example',
-    appId: 'app.capgo.audiorecorder',
-    channelName: 'production',
-    channelId: 38925,
-    version: '1.0.0',
-    webUrl: 'https://c38925-app-0capgo-0audiorecorder.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-audiorecorder-production-38925.webp',
-  },
-  'capacitor-auto': {
-    appName: 'Plugin Template Example',
-    appId: 'app.capgo.auto.example',
-    channelName: 'production',
-    channelId: 38873,
-    version: '1.0.0',
-    webUrl: 'https://c38873-app-0capgo-0auto-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-auto-example-production-38873.webp',
-  },
-  'capacitor-autofill-save-password': {
-    appName: 'Autofill Save Password Example',
-    appId: 'app.capgo.autofill.save.password.example',
-    channelName: 'production',
-    channelId: 38937,
-    version: '1.0.0',
-    webUrl: 'https://c38937-app-0capgo-0autofill-0save-0password-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-autofill-save-password-example-production-38937.webp',
-  },
-  'capacitor-background-geolocation': {
-    appName: 'Background Geolocation Example',
-    appId: 'com.capgo.backgroundgeolocation.example',
-    channelName: 'production',
-    channelId: 38888,
-    version: '1.0.0',
-    webUrl: 'https://c38888-com-0capgo-0backgroundgeolocation-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/com-capgo-backgroundgeolocation-example-production-38888.webp',
-  },
-  'capacitor-background-task': {
-    appName: 'Background Task Example',
-    appId: 'app.capgo.backgroundtask.example',
-    channelName: 'production',
-    channelId: 38864,
-    version: '1.0.0',
-    webUrl: 'https://c38864-app-0capgo-0backgroundtask-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-backgroundtask-example-production-38864.webp',
-  },
-  'capacitor-barometer': {
-    appName: 'Barometer Example',
-    appId: 'app.capgo.barometer',
-    channelName: 'production',
-    channelId: 38939,
-    version: '1.0.0',
-    webUrl: 'https://c38939-app-0capgo-0barometer.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-barometer-production-38939.webp',
-  },
-  'capacitor-bluetooth-low-energy': {
-    appName: 'BLE Example',
-    appId: 'ee.forgr.plugin.bluetoothlowenergy.example',
-    channelName: 'production',
-    channelId: 38887,
-    version: '0.0.0',
-    webUrl: 'https://c38887-ee-0forgr-0plugin-0bluetoothlowenergy-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/ee-forgr-plugin-bluetoothlowenergy-example-production-38887.webp',
-  },
-  'capacitor-brightness': {
-    appName: 'Brightness Example',
-    appId: 'app.capgo.brightness',
-    channelName: 'production',
-    channelId: 38979,
-    version: '1.0.0',
-    webUrl: 'https://c38979-app-0capgo-0brightness.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-brightness-production-38979.webp',
-  },
-  'capacitor-calendar': {
-    appName: 'Capgo Calendar',
-    appId: 'app.capgo.calendar.example',
-    channelName: 'production',
-    channelId: 37574,
-    version: '8.0.7',
-    webUrl: 'https://c37574-app-0capgo-0calendar-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-calendar-example-production-37574.webp',
-  },
-  'capacitor-camera-preview': {
-    appName: 'Camera Preview Example',
-    appId: 'app.capgo.camera.preview',
-    channelName: 'production',
-    channelId: 38918,
-    version: '0.0.1',
-    webUrl: 'https://c38918-app-0capgo-0camera-0preview.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-camera-preview-production-38918.webp',
-  },
-  'capacitor-compass': {
-    appName: 'Compass Example',
-    appId: 'app.capgo.compass',
-    channelName: 'production',
-    channelId: 38879,
-    version: '1.0.0',
-    webUrl: 'https://c38879-app-0capgo-0compass.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-compass-production-38879.webp',
-  },
-  'capacitor-contacts': {
-    appName: 'Contacts Example',
-    appId: 'app.capgo.contacts',
-    channelName: 'production',
-    channelId: 38886,
-    version: '1.0.0',
-    webUrl: 'https://c38886-app-0capgo-0contacts.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-contacts-production-38886.webp',
-  },
-  'capacitor-contentsquare': {
-    appName: 'Contentsquare Example',
-    appId: 'app.capgo.contentsquare.example',
-    channelName: 'production',
-    channelId: 38912,
-    version: '1.0.0',
-    webUrl: 'https://c38912-app-0capgo-0contentsquare-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-contentsquare-example-production-38912.webp',
-  },
-  'capacitor-crisp': {
-    appName: 'Crisp Example',
-    appId: 'app.capgo.crisp',
-    channelName: 'production',
-    channelId: 38862,
-    version: '1.0.0',
-    webUrl: 'https://c38862-app-0capgo-0crisp.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-crisp-production-38862.webp',
-  },
-  'capacitor-data-storage-sqlite': {
-    appName: 'Data Storage SQLite Example',
-    appId: 'app.capgo.data.storage.sqlite',
-    channelName: 'production',
-    channelId: 38931,
-    version: '1.0.0',
-    webUrl: 'https://c38931-app-0capgo-0data-0storage-0sqlite.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-data-storage-sqlite-production-38931.webp',
-  },
-  'capacitor-date-picker': {
-    appName: 'Capgo Date Picker',
-    appId: 'app.capgo.datepicker.example',
-    channelName: 'production',
-    channelId: 37575,
-    version: '8.1.1-22dd072',
-    webUrl: 'https://c37575-app-0capgo-0datepicker-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-datepicker-example-production-37575.webp',
-  },
-  'capacitor-device-info': {
-    appName: 'Device Info Example',
-    appId: 'app.capgo.deviceinfo.example',
-    channelName: 'production',
-    channelId: 38927,
-    version: '1.0.0',
-    webUrl: 'https://c38927-app-0capgo-0deviceinfo-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-deviceinfo-example-production-38927.webp',
-  },
-  'capacitor-document-scanner': {
-    appName: 'Document Scanner Example',
-    appId: 'app.capgo.document.scanner',
-    channelName: 'production',
-    channelId: 38891,
-    version: '1.0.0',
-    webUrl: 'https://c38891-app-0capgo-0document-0scanner.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-document-scanner-production-38891.webp',
-  },
-  'capacitor-downloader': {
-    appName: 'Downloader Example',
-    appId: 'app.capgo.downloader',
-    channelName: 'production',
-    channelId: 38936,
-    version: '1.0.0',
-    webUrl: 'https://c38936-app-0capgo-0downloader.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-downloader-production-38936.webp',
-  },
-  'capacitor-env': {
-    appName: 'Environment Example',
-    appId: 'app.capgo.env',
-    channelName: 'production',
-    channelId: 38942,
-    version: '1.0.0',
-    webUrl: 'https://c38942-app-0capgo-0env.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-env-production-38942.webp',
-  },
-  'capacitor-facebook-analytics': {
-    appName: 'Facebook Analytics Example',
-    appId: 'app.capgo.facebookanalytics.example',
-    channelName: 'production',
-    channelId: 38857,
-    version: '1.0.0',
-    webUrl: 'https://c38857-app-0capgo-0facebookanalytics-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-facebookanalytics-example-production-38857.webp',
-  },
-  'capacitor-fast-sql': {
-    appName: 'FastSQL Example',
-    appId: 'com.capgo.fastsql.example',
-    channelName: 'production',
-    channelId: 38907,
-    version: '1.0.0',
-    webUrl: 'https://c38907-com-0capgo-0fastsql-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/com-capgo-fastsql-example-production-38907.webp',
-  },
-  'capacitor-file': {
-    appName: 'File Example',
-    appId: 'app.capgo.file',
-    channelName: 'production',
-    channelId: 38831,
-    version: '1.0.0',
-    webUrl: 'https://c38831-app-0capgo-0file.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-file-production-38831.webp',
-  },
-  'capacitor-file-compressor': {
-    appName: 'File Compressor Example',
-    appId: 'app.capgo.filecompressor',
-    channelName: 'production',
-    channelId: 38922,
-    version: '1.0.0',
-    webUrl: 'https://c38922-app-0capgo-0filecompressor.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-filecompressor-production-38922.webp',
-  },
-  'capacitor-file-sharer': {
-    appName: 'Plugin Template Example',
-    appId: 'app.capgo.filesharer.example',
-    channelName: 'production',
-    channelId: 38905,
-    version: '1.0.0',
-    webUrl: 'https://c38905-app-0capgo-0filesharer-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-filesharer-example-production-38905.webp',
-  },
-  'capacitor-flash': {
-    appName: 'Flash Example',
-    appId: 'app.capgo.flash',
-    channelName: 'production',
-    channelId: 38896,
-    version: '1.0.0',
-    webUrl: 'https://c38896-app-0capgo-0flash.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-flash-production-38896.webp',
-  },
-  'capacitor-gtm': {
-    appName: 'GTM Example',
-    appId: 'app.capgo.gtm',
-    channelName: 'production',
-    channelId: 38855,
-    version: '1.0.0',
-    webUrl: 'https://c38855-app-0capgo-0gtm.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-gtm-production-38855.webp',
-  },
-  'capacitor-health': {
-    appName: 'Health Example',
-    appId: 'app.capgo.health',
-    channelName: 'production',
-    channelId: 38903,
-    version: '1.0.0',
-    webUrl: 'https://c38903-app-0capgo-0health.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-health-production-38903.webp',
-  },
-  'capacitor-home-indicator': {
-    appName: 'Home Indicator Example',
-    appId: 'app.capgo.home.indicator.example',
-    channelName: 'production',
-    channelId: 38900,
-    version: '1.0.0',
-    webUrl: 'https://c38900-app-0capgo-0home-0indicator-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-home-indicator-example-production-38900.webp',
-  },
-  'capacitor-ibeacon': {
-    appName: 'iBeacon Example',
-    appId: 'app.capgo.ibeacon',
-    channelName: 'production',
-    channelId: 38851,
-    version: '1.0.0',
-    webUrl: 'https://c38851-app-0capgo-0ibeacon.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-ibeacon-production-38851.webp',
-  },
-  'capacitor-inappbrowser': {
-    appName: 'Inappbrowser Example',
-    appId: 'app.capgo.inappbrowser',
-    channelName: 'production',
-    channelId: 38930,
-    version: '1.0.0',
-    webUrl: 'https://c38930-app-0capgo-0inappbrowser.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-inappbrowser-production-38930.webp',
-  },
-  'capacitor-incoming-call-kit': {
-    appName: 'Privacy Screen Example',
-    appId: 'app.capgo.incomingcallkit.example',
-    channelName: 'production',
-    channelId: 38841,
-    version: '1.0.0',
-    webUrl: 'https://c38841-app-0capgo-0incomingcallkit-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-incomingcallkit-example-production-38841.webp',
-  },
-  'capacitor-install-referrer': {
-    appName: 'Install Referrer Example',
-    appId: 'app.capgo.installreferrer.example',
-    channelName: 'production',
-    channelId: 38832,
-    version: '1.0.0',
-    webUrl: 'https://c38832-app-0capgo-0installreferrer-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-installreferrer-example-production-38832.webp',
-  },
-  'capacitor-intent-launcher': {
-    appName: 'Intent Launcher Example',
-    appId: 'app.capgo.intentlauncher',
-    channelName: 'production',
-    channelId: 38904,
-    version: '1.0.0',
-    webUrl: 'https://c38904-app-0capgo-0intentlauncher.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-intentlauncher-production-38904.webp',
-  },
-  'capacitor-is-root': {
-    appName: 'Is Root Example',
-    appId: 'app.capgo.is.root',
-    channelName: 'production',
-    channelId: 38902,
-    version: '1.0.0',
-    webUrl: 'https://c38902-app-0capgo-0is-0root.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-is-root-production-38902.webp',
-  },
-  'capacitor-ivs-player': {
-    appName: 'IVS Player Example',
-    appId: 'app.capgo.ivs.player.example',
-    channelName: 'production',
-    channelId: 38871,
-    version: '1.0.0',
-    webUrl: 'https://c38871-app-0capgo-0ivs-0player-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-ivs-player-example-production-38871.webp',
-  },
-  'capacitor-jw-player': {
-    appName: 'Jw Player Example',
-    appId: 'app.capgo.jw.player',
-    channelName: 'production',
-    channelId: 38964,
-    version: '1.0.0',
-    webUrl: 'https://c38964-app-0capgo-0jw-0player.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-jw-player-production-38964.webp',
-  },
-  'capacitor-keep-awake': {
-    appName: 'Keep Awake Example',
-    appId: 'app.capgo.keepawake',
-    channelName: 'production',
-    channelId: 38919,
-    version: '1.0.0',
-    webUrl: 'https://c38919-app-0capgo-0keepawake.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-keepawake-production-38919.webp',
-  },
-  'capacitor-launch-navigator': {
-    appName: 'Launch Navigator Example',
-    appId: 'app.capgo.launch.navigator',
-    channelName: 'production',
-    channelId: 38880,
-    version: '1.0.0',
-    webUrl: 'https://c38880-app-0capgo-0launch-0navigator.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-launch-navigator-production-38880.webp',
-  },
-  'capacitor-light-sensor': {
-    appName: 'Light Sensor Example',
-    appId: 'app.capgo.lightsensor',
-    channelName: 'production',
-    channelId: 38882,
-    version: '0.0.1',
-    webUrl: 'https://c38882-app-0capgo-0lightsensor.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-lightsensor-production-38882.webp',
-  },
-  'capacitor-live-reload': {
-    appName: 'Live Reload Example',
-    appId: 'app.capgo.livereload',
-    channelName: 'production',
-    channelId: 38897,
-    version: '1.0.0',
-    webUrl: 'https://c38897-app-0capgo-0livereload.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-livereload-production-38897.webp',
-  },
-  'capacitor-llm': {
-    appName: 'LLM Example',
-    appId: 'app.capgo.llm',
-    channelName: 'production',
-    channelId: 38940,
-    version: '0.0.1',
-    webUrl: 'https://c38940-app-0capgo-0llm.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-llm-production-38940.webp',
-  },
-  'capacitor-media-session': {
-    appName: 'Media Session Example',
-    appId: 'app.capgo.media.session',
-    channelName: 'production',
-    channelId: 38941,
-    version: '1.0.0',
-    webUrl: 'https://c38941-app-0capgo-0media-0session.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-media-session-production-38941.webp',
-  },
-  'capacitor-mqtt': {
-    appName: 'MQTT Example',
-    appId: 'app.capgo.mqtt.example',
-    channelName: 'production',
-    channelId: 38858,
-    version: '1.0.0',
-    webUrl: 'https://c38858-app-0capgo-0mqtt-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-mqtt-example-production-38858.webp',
-  },
-  'capacitor-mute': {
-    appName: 'Mute Example',
-    appId: 'app.capgo.mute',
-    channelName: 'production',
-    channelId: 38833,
-    version: '1.0.0',
-    webUrl: 'https://c38833-app-0capgo-0mute.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-mute-production-38833.webp',
-  },
-  'capacitor-mux-player': {
-    appName: 'Mux Player Example',
-    appId: 'app.capgo.mux.player',
-    channelName: 'production',
-    channelId: 38898,
-    version: '1.0.0',
-    webUrl: 'https://c38898-app-0capgo-0mux-0player.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-mux-player-production-38898.webp',
-  },
-  'capacitor-native-biometric': {
-    appName: 'example-app',
-    appId: 'app.capgo.native.biometric',
-    channelName: 'production',
-    channelId: 38893,
-    version: '1.0.0',
-    webUrl: 'https://c38893-app-0capgo-0native-0biometric.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-native-biometric-production-38893.webp',
-  },
-  'capacitor-native-loader': {
-    appName: 'Native Loader Example',
-    appId: 'app.capgo.nativeloader.example',
-    channelName: 'production',
-    channelId: 38926,
-    version: '1.0.0',
-    webUrl: 'https://c38926-app-0capgo-0nativeloader-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-nativeloader-example-production-38926.webp',
-  },
-  'capacitor-native-market': {
-    appName: 'Native Market Example',
-    appId: 'app.capgo.nativemarket',
-    channelName: 'production',
-    channelId: 38911,
-    version: '1.0.0',
-    webUrl: 'https://c38911-app-0capgo-0nativemarket.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-nativemarket-production-38911.webp',
-  },
-  'capacitor-native-navigation': {
-    appName: 'Native Navigation Example',
-    appId: 'app.capgo.native.navigation.example',
-    channelName: 'production',
-    channelId: 38546,
-    version: '1.0.0',
-    webUrl: 'https://c38546-app-0capgo-0native-0navigation-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-native-navigation-example-production-38546.webp',
-  },
-  'capacitor-native-purchases': {
-    appName: 'Native Purchases Example',
-    appId: 'app.capgo.nativepurchases',
-    channelName: 'production',
-    channelId: 38889,
-    version: '1.0.0',
-    webUrl: 'https://c38889-app-0capgo-0nativepurchases.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-nativepurchases-production-38889.webp',
-  },
-  'capacitor-nativegeocoder': {
-    appName: 'Native Geocoder Example',
-    appId: 'app.capgo.nativegeocoder',
-    channelName: 'production',
-    channelId: 38842,
-    version: '1.0.0',
-    webUrl: 'https://c38842-app-0capgo-0nativegeocoder.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-nativegeocoder-production-38842.webp',
-  },
-  'capacitor-navigation-bar': {
-    appName: 'Navigation Bar Example',
-    appId: 'app.capgo.navigation.bar',
-    channelName: 'production',
-    channelId: 38854,
-    version: '0.0.1',
-    webUrl: 'https://c38854-app-0capgo-0navigation-0bar.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-navigation-bar-production-38854.webp',
-  },
-  'capacitor-network-diagnostics': {
-    appName: 'Network Diagnostics Example',
-    appId: 'app.capgo.networkdiagnostics.example',
-    channelName: 'production',
-    channelId: 38869,
-    version: '1.0.0',
-    webUrl: 'https://c38869-app-0capgo-0networkdiagnostics-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-networkdiagnostics-example-production-38869.webp',
-  },
-  'capacitor-nfc': {
-    appName: 'NFC Example',
-    appId: 'app.capgo.nfc.example',
-    channelName: 'production',
-    channelId: 38932,
-    version: '1.0.0',
-    webUrl: 'https://c38932-app-0capgo-0nfc-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-nfc-example-production-38932.webp',
-  },
-  'capacitor-passkey': {
-    appName: 'Capgo Passkey Demo',
-    appId: 'app.capgo.passkey.example',
-    channelName: 'production',
-    channelId: 38853,
-    version: '1.0.0',
-    webUrl: 'https://c38853-app-0capgo-0passkey-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-passkey-example-production-38853.webp',
-  },
-  'capacitor-pay': {
-    appName: 'Pay Example',
-    appId: 'app.capgo.pay',
-    channelName: 'production',
-    channelId: 38870,
-    version: '1.0.0',
-    webUrl: 'https://c38870-app-0capgo-0pay.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-pay-production-38870.webp',
-  },
-  'capacitor-pdf-generator': {
-    appName: 'PDF Generator Example',
-    appId: 'app.capgo.pdfgenerator',
-    channelName: 'production',
-    channelId: 38861,
-    version: '1.0.0',
-    webUrl: 'https://c38861-app-0capgo-0pdfgenerator.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-pdfgenerator-production-38861.webp',
-  },
-  'capacitor-pedometer': {
-    appName: 'Pedometer Example',
-    appId: 'app.capgo.pedometer',
-    channelName: 'production',
-    channelId: 38909,
-    version: '1.0.0',
-    webUrl: 'https://c38909-app-0capgo-0pedometer.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-pedometer-production-38909.webp',
-  },
-  'capacitor-persistent-account': {
-    appName: 'Persistent Account Example',
-    appId: 'app.capgo.persistent.account',
-    channelName: 'production',
-    channelId: 38849,
-    version: '1.0.0',
-    webUrl: 'https://c38849-app-0capgo-0persistent-0account.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-persistent-account-production-38849.webp',
-  },
-  'capacitor-photo-library': {
-    appName: 'Photo Library Example',
-    appId: 'app.capgo.photo.library',
-    channelName: 'production',
-    channelId: 38894,
-    version: '1.0.0',
-    webUrl: 'https://c38894-app-0capgo-0photo-0library.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-photo-library-production-38894.webp',
-  },
-  'capacitor-pretty-toast': {
-    appName: 'Pretty Toast',
-    appId: 'app.capgo.pretty.toast',
-    channelName: 'production',
-    channelId: 37182,
-    version: '8.1.10',
-    webUrl: 'https://c37182-app-0capgo-0pretty-0toast.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-pretty-toast-production-37182.webp',
-  },
-  'capacitor-printer': {
-    appName: 'CapacitorPrinterExample',
-    appId: 'com.capgo.printer.example',
-    channelName: 'production',
-    channelId: 38885,
-    version: '1.0.0',
-    webUrl: 'https://c38885-com-0capgo-0printer-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/com-capgo-printer-example-production-38885.webp',
-  },
-  'capacitor-privacy-screen': {
-    appName: 'Privacy Screen Example',
-    appId: 'app.capgo.privacyscreen.example',
-    channelName: 'production',
-    channelId: 38916,
-    version: '1.0.0',
-    webUrl: 'https://c38916-app-0capgo-0privacyscreen-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-privacyscreen-example-production-38916.webp',
-  },
-  'capacitor-proximity': {
-    appName: 'Capgo Proximity Demo',
-    appId: 'app.capgo.proximity.example',
-    channelName: 'production',
-    channelId: 38901,
-    version: '1.0.0',
-    webUrl: 'https://c38901-app-0capgo-0proximity-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-proximity-example-production-38901.webp',
-  },
-  'capacitor-realtimekit': {
-    appName: 'RealtimeKit Example',
-    appId: 'app.capgo.realtimekit',
-    channelName: 'production',
-    channelId: 38848,
-    version: '1.0.0',
-    webUrl: 'https://c38848-app-0capgo-0realtimekit.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-realtimekit-production-38848.webp',
-  },
-  'capacitor-recaptcha': {
-    appName: 'Recaptcha Example',
-    appId: 'app.capgo.recaptcha.example',
-    channelName: 'production',
-    channelId: 38881,
-    version: '1.0.0',
-    webUrl: 'https://c38881-app-0capgo-0recaptcha-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-recaptcha-example-production-38881.webp',
-  },
-  'capacitor-ricoh360-camera-plugin': {
-    appName: 'Ricoh360 Camera Plugin Example',
-    appId: 'app.capgo.ricoh360.camera.plugin',
-    channelName: 'production',
-    channelId: 38868,
-    version: '1.0.0',
-    webUrl: 'https://c38868-app-0capgo-0ricoh360-0camera-0plugin.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-ricoh360-camera-plugin-production-38868.webp',
-  },
-  'capacitor-rudderstack': {
-    appName: 'RudderStack Example',
-    appId: 'app.capgo.rudderstack.example',
-    channelName: 'production',
-    channelId: 38906,
-    version: '1.0.0',
-    webUrl: 'https://c38906-app-0capgo-0rudderstack-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-rudderstack-example-production-38906.webp',
-  },
-  'capacitor-screen-orientation': {
-    appName: 'Screen Orientation Example',
-    appId: 'app.capgo.screenorientation',
-    channelName: 'production',
-    channelId: 38934,
-    version: '1.0.0',
-    webUrl: 'https://c38934-app-0capgo-0screenorientation.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-screenorientation-production-38934.webp',
-  },
-  'capacitor-screen-recorder': {
-    appName: 'Screen Recorder Example',
-    appId: 'app.capgo.screen.recorder',
-    channelName: 'production',
-    channelId: 38884,
-    version: '1.0.0',
-    webUrl: 'https://c38884-app-0capgo-0screen-0recorder.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-screen-recorder-production-38884.webp',
-  },
-  'capacitor-shake': {
-    appName: 'Shake Example',
-    appId: 'app.capgo.shake',
-    channelName: 'production',
-    channelId: 38938,
-    version: '1.0.0',
-    webUrl: 'https://c38938-app-0capgo-0shake.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-shake-production-38938.webp',
-  },
-  'capacitor-share-target': {
-    appName: 'Share Target Example',
-    appId: 'app.capgo.sharetarget',
-    channelName: 'production',
-    channelId: 38928,
-    version: '1.0.0',
-    webUrl: 'https://c38928-app-0capgo-0sharetarget.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-sharetarget-production-38928.webp',
-  },
-  'capacitor-sheets': {
-    appName: 'Capgo Sheets',
-    appId: 'app.capgo.capacitor.sheets',
-    channelName: 'production',
-    channelId: 37983,
-    version: '8.0.17',
-    webUrl: 'https://c37983-app-0capgo-0capacitor-0sheets.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-capacitor-sheets-production-37983.webp',
-  },
-  'capacitor-sim': {
-    appName: 'Sim Example',
-    appId: 'app.capgo.sim',
-    channelName: 'production',
-    channelId: 38910,
-    version: '1.0.0',
-    webUrl: 'https://c38910-app-0capgo-0sim.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-sim-production-38910.webp',
-  },
-  'capacitor-speech-recognition': {
-    appName: 'Speech Recognition Example',
-    appId: 'app.capgo.speechrecognition',
-    channelName: 'production',
-    channelId: 38935,
-    version: '1.0.0',
-    webUrl: 'https://c38935-app-0capgo-0speechrecognition.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-speechrecognition-production-38935.webp',
-  },
-  'capacitor-speech-synthesis': {
-    appName: 'Speech Synthesis Example',
-    appId: 'app.capgo.speechsynthesis',
-    channelName: 'production',
-    channelId: 38872,
-    version: '1.0.0',
-    webUrl: 'https://c38872-app-0capgo-0speechsynthesis.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-speechsynthesis-production-38872.webp',
-  },
-  'capacitor-ssl-pinning': {
-    appName: 'SSL Pinning Example',
-    appId: 'app.capgo.sslpinning.example',
-    channelName: 'production',
-    channelId: 38923,
-    version: '1.0.0',
-    webUrl: 'https://c38923-app-0capgo-0sslpinning-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-sslpinning-example-production-38923.webp',
-  },
-  'capacitor-supabase': {
-    appName: 'Supabase Example',
-    appId: 'app.capgo.supabase.example',
-    channelName: 'production',
-    channelId: 38878,
-    version: '1.0.0',
-    webUrl: 'https://c38878-app-0capgo-0supabase-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-supabase-example-production-38878.webp',
-  },
-  'capacitor-textinteraction': {
-    appName: 'Text Interaction Example',
-    appId: 'app.capgo.textinteraction',
-    channelName: 'production',
-    channelId: 38917,
-    version: '1.0.0',
-    webUrl: 'https://c38917-app-0capgo-0textinteraction.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-textinteraction-production-38917.webp',
-  },
-  'capacitor-transitions': {
-    appName: 'Capgo Transitions',
-    appId: 'app.capgo.capacitor.transitions',
-    channelName: 'production',
-    channelId: 37981,
-    version: '8.1.9',
-    webUrl: 'https://c37981-app-0capgo-0capacitor-0transitions.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-capacitor-transitions-production-37981.webp',
-  },
-  'capacitor-twilio-voice': {
-    appName: 'Twilio Voice Example',
-    appId: 'app.capgo.twilio.voice',
-    channelName: 'production',
-    channelId: 38883,
-    version: '1.0.0',
-    webUrl: 'https://c38883-app-0capgo-0twilio-0voice.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-twilio-voice-production-38883.webp',
-  },
-  'capacitor-updater': {
-    appName: 'Updater Example',
-    appId: 'app.capgo.updater',
-    channelName: 'production',
-    channelId: 39087,
-    version: '1.0.0',
-    webUrl: 'https://c39087-app-0capgo-0updater.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-updater-production-39087.webp',
-  },
-  'capacitor-uploader': {
-    appName: 'Uploader Example',
-    appId: 'app.capgo.uploader',
-    channelName: 'production',
-    channelId: 38929,
-    version: '1.0.0',
-    webUrl: 'https://c38929-app-0capgo-0uploader.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-uploader-production-38929.webp',
-  },
-  'capacitor-verisoul': {
-    appName: 'Verisoul Example',
-    appId: 'app.capgo.verisoul.example',
-    channelName: 'production',
-    channelId: 38899,
-    version: '1.0.0',
-    webUrl: 'https://c38899-app-0capgo-0verisoul-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-verisoul-example-production-38899.webp',
-  },
-  'capacitor-video-player': {
-    appName: 'Video Player Example',
-    appId: 'app.capgo.video.player',
-    channelName: 'production',
-    channelId: 38877,
-    version: '1.0.0',
-    webUrl: 'https://c38877-app-0capgo-0video-0player.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-video-player-production-38877.webp',
-  },
-  'capacitor-video-thumbnails': {
-    appName: 'Video Thumbnails Example',
-    appId: 'app.capgo.videothumbnails',
-    channelName: 'production',
-    channelId: 38913,
-    version: '1.0.0',
-    webUrl: 'https://c38913-app-0capgo-0videothumbnails.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-videothumbnails-production-38913.webp',
-  },
-  'capacitor-volume-buttons': {
-    appName: 'Volume Buttons Example',
-    appId: 'app.capgo.volume.buttons',
-    channelName: 'production',
-    channelId: 38847,
-    version: '1.0.0',
-    webUrl: 'https://c38847-app-0capgo-0volume-0buttons.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-volume-buttons-production-38847.webp',
-  },
-  'capacitor-watch': {
-    appName: 'Watch Example',
-    appId: 'app.capgo.watch',
-    channelName: 'production',
-    channelId: 38892,
-    version: '1.0.0',
-    webUrl: 'https://c38892-app-0capgo-0watch.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-watch-production-38892.webp',
-  },
-  'capacitor-webview-crash': {
-    appName: 'WebView Crash Demo',
-    appId: 'app.capgo.webviewcrash.example',
-    channelName: 'production',
-    channelId: 38890,
-    version: '1.0.0',
-    webUrl: 'https://c38890-app-0capgo-0webviewcrash-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-webviewcrash-example-production-38890.webp',
-  },
-  'capacitor-webview-guardian': {
-    appName: 'Webview Guardian Example',
-    appId: 'app.capgo.webviewguardian.example',
-    channelName: 'production',
-    channelId: 38924,
-    version: '1.0.0',
-    webUrl: 'https://c38924-app-0capgo-0webviewguardian-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-webviewguardian-example-production-38924.webp',
-  },
-  'capacitor-webview-version-checker': {
-    appName: 'WebView Version Checker Example',
-    appId: 'app.capgo.webviewversionchecker.example',
-    channelName: 'production',
-    channelId: 38948,
-    version: '1.0.0',
-    webUrl: 'https://c38948-app-0capgo-0webviewversionchecker-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-webviewversionchecker-example-production-38948.webp',
-  },
-  'capacitor-wechat': {
-    appName: 'WeChat Example',
-    appId: 'app.capgo.wechat',
-    channelName: 'production',
-    channelId: 38875,
-    version: '1.0.0',
-    webUrl: 'https://c38875-app-0capgo-0wechat.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-wechat-production-38875.webp',
-  },
-  'capacitor-widget-kit': {
-    appName: 'Widget Kit Example',
-    appId: 'app.capgo.widgetkit.exampleapp',
-    channelName: 'production',
-    channelId: 38836,
-    version: '1.0.0',
-    webUrl: 'https://c38836-app-0capgo-0widgetkit-0exampleapp.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-widgetkit-exampleapp-production-38836.webp',
-  },
-  'capacitor-wifi': {
-    appName: 'WiFi Example',
-    appId: 'app.capgo.wifi',
-    channelName: 'production',
-    channelId: 38856,
-    version: '1.0.0',
-    webUrl: 'https://c38856-app-0capgo-0wifi.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-wifi-production-38856.webp',
-  },
-  'capacitor-youtube-player': {
-    appName: 'YouTube Player Example',
-    appId: 'app.capgo.youtube.player',
-    channelName: 'production',
-    channelId: 38846,
-    version: '1.0.0',
-    webUrl: 'https://c38846-app-0capgo-0youtube-0player.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-youtube-player-production-38846.webp',
-  },
-  'capacitor-zebra-datawedge': {
-    appName: 'Zebra DataWedge Example',
-    appId: 'app.capgo.zebradatawedge.example',
-    channelName: 'production',
-    channelId: 38914,
-    version: '1.0.0',
-    webUrl: 'https://c38914-app-0capgo-0zebradatawedge-0example.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-zebradatawedge-example-production-38914.webp',
-  },
-  'capacitor-zip': {
-    appName: 'Zip Example',
-    appId: 'app.capgo.zip',
-    channelName: 'production',
-    channelId: 38863,
-    version: '1.0.0',
-    webUrl: 'https://c38863-app-0capgo-0zip.preview.capgo.app/',
-    qr: '/plugins/preview-qrs/app-capgo-zip-production-38863.webp',
-  },
-}
-
-export const getPluginPreviewQr = (slug?: string): PluginPreviewQr | undefined => (slug ? pluginPreviewQrs[slug] : undefined)
+export { getPluginPreviewQr, pluginPreviewQrs, type PluginPreviewQr } from '../../../shared/pluginPreviewQrs'
